chore: move model type code from `tracer/mod.rs` to `tracer/model.rs`
